+++ Archard/src/main/java/com/archidelion/archard/dao/CharactDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.archidelion.archard.bean.charact.Charact;
 
-public interface CharactDao extends JpaRepository<Charact, Integer> {
-	Charact getCharacterById(int id);
+public interface CharactDao extends JpaRepository<Charact, Long> {
+	Charact getCharacterById(Long id);
 }
