+++ Archard/src/main/java/com/archidelion/archard/service/CharactService.java
@@ -5,5 +5,5 @@ import com.archidelion.archard.bean.charact.Charact;
 public interface CharactService {
 	Charact addCharacter(String name, int userId);
 
-	Charact getCharacterById(int id);
+	Charact getCharacterById(Long id);
 }
